requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.8
cmake==3.30.0
colorama==0.4.6
comtypes==1.4.4
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
distro==1.9.0
fastapi==0.115.12
filelock==3.16.1
filterpy==1.4.5
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.55.3
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
groq==0.22.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.7
imageio==2.36.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lap==0.5.12
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.22.0
natsort==8.4.0
networkx==3.4.2
numpy==2.0.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.2
pillow==11.1.0
protobuf==5.29.3
pyarrow==19.0.0
pydantic==2.11.2
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyttsx3==2.90
pytz==2024.1
PyYAML==6.0.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.3
scikit-image==0.25.0
scipy==1.15.1
setuptools==75.8.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
tifffile==2025.1.10
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.34.0
watchdog==6.0.0
Werkzeug==3.1.3